+++ hackathon/src/main/java/com/hackathon/service/LoginService.java
@@ -1,6 +1,6 @@
 package com.hackathon.service;
 
-import org.springframework.stereotype.Service;
+import org.springframework.stereotype.Component;
 
 import com.hackathon.exceptions.AuthException;
 import com.hackathon.models.UserRequest;
@@ -10,8 +10,8 @@ import com.hackathon.models.UserRequest;
  * @author kiranavk
  *
  */
-@Service
-public class LoginService {
+@Component
+public interface LoginService {
 	
 	/**
 	 * This method validates the user by verifying his username and password
@@ -19,18 +19,5 @@ public class LoginService {
 	 * @return String LoginResult
 	 * @throws AuthException
 	 */
-	public String validateUser(UserRequest user) throws AuthException {
-		String username = user.getUserName();
-		String password = user.getPassword();
-		if(username == null || "".equals(username)) {
-			throw new AuthException("Username should not be empty.");
-		}
-		if(password == null || "".equals(password)) {
-			throw new AuthException("Password should not be empty.");
-		}
-		if(!"kiran".equalsIgnoreCase(username) || !"avk".equalsIgnoreCase(password)) {
-			throw new AuthException("Username or Password is invalid.");
-		}
-		return "success";
-	}
+	public String validateUser(UserRequest user) throws AuthException;
 }
